feature/admin-manager 슈퍼유저 삭제 방지 로직 추가

+++ src/main/java/com/swj9707/twittercloneadmin/user/service/AdminUserServiceImpl.java
@@ -62,7 +62,13 @@ public class AdminUserServiceImpl implements AdminUserService {
 
     @Override
     public void delete(Long userID) {
-        adminUserRepository.deleteById(userID);
+        AdminUserEntity entity = adminUserRepository.findById(userID)
+                        .orElseThrow(() -> new CustomException(ErrorCode.USER_NOTFOUND));
+        if(entity.getRole().equals(Role.SUPERADMIN)){
+            throw new CustomException(ErrorCode.BAD_REQUEST);
+        } else {
+            adminUserRepository.deleteById(userID);
+        }
     }
 
 }
